requirements.txt
fastapi
uvicorn
beanie
motor
pydantic
passlib
python-jose
pyotp
cryptography
dotenv
pydantic[email]
pydantic_settings
fastapi_mail
python-multipart
shapely
httpx
pandas
bcrypt
reportlab
